genie-ai-vibe/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.1

# AI and LLM dependencies
google-generativeai==0.3.2

# HTTP client
httpx==0.26.0
requests==2.31.0

# Data validation and parsing
pydantic==2.5.3

# CORS middleware (already included in FastAPI but explicit for clarity)
python-jose[cryptography]==3.3.0

# Database (SQLite is built-in, but adding for completeness)
# If you decide to use PostgreSQL or other databases in the future:
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.25

# Development and testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.2

# Optional: Logging and monitoring
structlog==23.2.0 

# Computer Vision and Face Recognition
opencv-python==4.10.0.84
mediapipe==0.10.8
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.0.1